added react-router-dom, added proper routes to App.js, modified connection.js (db name), modified server (previous to queries and mutation built)
