+++ server/routes/index.js
@@ -1,0 +1,6 @@
+const router = require('express').Router();
+const apiRoutes = require('./api');
+
+router.use('/api', apiRoutes);
+
+module.exports = router;
