+++ server/server.js
@@ -1,41 +1,58 @@
 const express = require('express');
-const { ApolloServer } = require('apollo-server-express');
+// Uncomment the following code once you have built the queries and mutations in the client folder
+// const { ApolloServer } = require('apollo-server-express');
 const path = require('path');
 
-const { typeDefs, resolvers } = require('./schemas');
+// Uncomment the following code once you have built the queries and mutations in the client folder
+// const { typeDefs, resolvers } = require('./schemas');
 const db = require('./config/connection');
+
+// Comment out this code once you have built out queries and mutations in the client folder
+const routes = require('./routes');
 
 const PORT = process.env.PORT || 3001;
 const app = express();
+// Uncomment the following code once you have built the queries and mutations in the client folder
+/*
 const server = new ApolloServer({
   typeDefs,
   resolvers,
 });
+*/
 
-app.use(express.urlencoded({ extended: false }));
+app.use(express.urlencoded({ extended: true }));
 app.use(express.json());
 
+// if we're in production, serve client/build as static assets
 if (process.env.NODE_ENV === 'production') {
   app.use(express.static(path.join(__dirname, '../client/build')));
 }
 
+// Comment out this code once you have built out queries and mutations in the client folder
+app.use(routes);
+
+// Uncomment the following code once you have built the queries and mutations in the client folder
+/* 
 app.get('/', (req, res) => {
   res.sendFile(path.join(__dirname, '../client/build/index.html'));
 });
 
-// Create a new instance of an Apollo server with the GraphQL schema
 const startApolloServer = async () => {
   await server.start();
   server.applyMiddleware({ app });
-  
+
   db.once('open', () => {
     app.listen(PORT, () => {
       console.log(`API server running on port ${PORT}!`);
       console.log(`Use GraphQL at http://localhost:${PORT}${server.graphqlPath}`);
-    })
-  })
-  };
-  
-  // Call the async function to start the server
-  startApolloServer();
+    });
+  });
+};
+*/
+
+// Comment out this code once you have built out queries and mutations in the client folder
+db.once('open', () => {
+  app.listen(PORT, () => console.log(`Now listening on localhost: ${PORT}`));
+});
+
